# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Awards Management API",
        "description": "API for managing awards, nominees, voting, and events",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Awards Management API Server"
        }
    ],
    "paths": {
...
        "/api/v1/registrations": {
            "post": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Create a new registration",
                "description": "Register for the awards event",
                "operationId": "createRegistration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone",
                                    "country",
                                    "city",
                                    "ticket_type",
                                    "event_date"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "organization": {
                                        "type": "string",
                                        "example": "Tech Corp"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "United States"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "dietary_requirements": {
                                        "type": "string",
                                        "example": "Vegetarian"
                                    },
                                    "ticket_type": {
                                        "type": "string",
                                        "enum": [
                                            "standard",
                                            "vip",
                                            "corporate"
                                        ],
                                        "example": "standard"
                                    },
                                    "event_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registration completed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "reference_number": {
                                                    "type": "string",
                                                    "example": "FGR24-123456"
                                                },
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "ticket_type": {
                                                    "type": "string",
                                                    "example": "standard"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 250
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "confirmed"
                                                },
                                                "event_date": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/registrations/lookup": {
            "post": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Look up registration",
                "description": "Find registration by reference number and email",
                "operationId": "lookupRegistration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reference_number",
                                    "email"
                                ],
                                "properties": {
                                    "reference_number": {
                                        "type": "string",
                                        "example": "FGR24-123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "reference_number": {
                                                    "type": "string"
                                                },
                                                "full_name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                },
                                                "organization": {
                                                    "type": "string"
                                                },
                                                "country": {
                                                    "type": "string"
                                                },
                                                "city": {
                                                    "type": "string"
                                                },
                                                "dietary_requirements": {
                                                    "type": "string"
                                                },
                                                "ticket_type": {
                                                    "type": "string"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "event_date": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registration not found"
                    }
                }
            }
        },
...
    },
...
}